api-python/requirements.txt
flask
flask-SQLAlchemy
diffusers
IPython.display
torch
gradio
